Extracted FullScreenAdViewController into separate file
